@@ -1,17 +1,16 @@
-
 {
     "name": "emilybache/parrot-refactoring-kata",
     "description": "Kata to practice Pol(l)ymorphism",
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0"
+        "php": "^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "phpstan/phpstan": "^0.12.25",
-        "phpstan/phpstan-phpunit": "^0.12.10",
-        "symplify/easy-coding-standard": "^9.0",
-        "symplify/phpstan-extensions": "^9.0"
+        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "symplify/easy-coding-standard": "^11.1",
+        "symplify/phpstan-extensions": "^11.1"
     },
     "autoload": {
         "psr-4": {
@@ -24,9 +23,6 @@
         }
     },
     "scripts": {
-        "checkcode": "phpcs src tests --standard=PSR12",
-        "fixcode": "phpcbf src tests --standard=PSR12",
-        "test": "phpunit",
         "tests": "phpunit",
         "test-coverage": "phpunit --coverage-html build/coverage",
         "check-cs": "ecs check --ansi",
