@@ -8,45 +8,28 @@ use Exception;
 
 class Parrot
 {
-    /**
-     * @var int ParrotTypeEnum
-     */
-    private $type;
-
-    /**
-     * @var int
-     */
-    private $numberOfCoconuts;
-
-    /**
-     * @var float
-     */
-    private $voltage;
-
-    /**
-     * @var bool
-     */
-    private $isNailed;
-
-    public function __construct(int $type, int $numberOfCoconuts, float $voltage, bool $isNailed)
-    {
-        $this->type = $type;
-        $this->numberOfCoconuts = $numberOfCoconuts;
-        $this->voltage = $voltage;
-        $this->isNailed = $isNailed;
+    public function __construct(
+        /**
+         * @var int ParrotTypeEnum
+         */
+        private int $type,
+        private int $numberOfCoconuts,
+        private float $voltage,
+        private bool $isNailed
+    ) {
     }
 
+    /**
+     * @throws Exception
+     */
     public function getSpeed(): float
     {
-        switch ($this->type) {
-            case ParrotTypeEnum::EUROPEAN:
-                return $this->getBaseSpeed();
-            case ParrotTypeEnum::AFRICAN:
-                return max(0, $this->getBaseSpeed() - $this->getLoadFactor() * $this->numberOfCoconuts);
-            case ParrotTypeEnum::NORWEGIAN_BLUE:
-                return $this->isNailed ? 0 : $this->getBaseSpeedWith($this->voltage);
-        }
-        throw new Exception('Should be unreachable');
+        return match ($this->type) {
+            ParrotTypeEnum::EUROPEAN => $this->getBaseSpeed(),
+            ParrotTypeEnum::AFRICAN => max(0, $this->getBaseSpeed() - $this->getLoadFactor() * $this->numberOfCoconuts),
+            ParrotTypeEnum::NORWEGIAN_BLUE => $this->isNailed ? 0 : $this->getBaseSpeedWith($this->voltage),
+            default => throw new Exception('Should be unreachable'),
+        };
     }
 
     private function getBaseSpeedWith(float $voltage): float
