@@ -3,25 +3,26 @@
 declare(strict_types=1);
 
 use PhpCsFixer\Fixer\ArrayNotation\ArraySyntaxFixer;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Symplify\EasyCodingStandard\ValueObject\Option;
+use PhpCsFixer\Fixer\Strict\DeclareStrictTypesFixer;
+use Symplify\EasyCodingStandard\Config\ECSConfig;
 use Symplify\EasyCodingStandard\ValueObject\Set\SetList;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $services = $containerConfigurator->services();
-    $services->set(ArraySyntaxFixer::class)
-        ->call('configure', [[
-            'syntax' => 'short',
-        ]]);
+// composer require --dev symplify/easy-coding-standard
+// vendor/bin/ecs init
 
-    $parameters = $containerConfigurator->parameters();
-    $parameters->set(Option::PATHS, [
+return static function (ECSConfig $ecsConfig): void {
+    $ecsConfig->paths([
         __DIR__ . '/src',
         __DIR__ . '/tests',
+        __DIR__ . '/ecs.php',  // check this file too!
     ]);
 
-    $parameters->set(Option::SETS, [
-//         run and fix, one by one
+    $ecsConfig->skip([
+        // skip specific rules
+    ]);
+
+    // run and fix, one by one
+    $ecsConfig->sets([
         SetList::SPACES,
         SetList::ARRAY,
         SetList::DOCBLOCK,
@@ -30,8 +31,17 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         SetList::CLEAN_CODE,
         SetList::STRICT,
         SetList::PSR_12,
-        SetList::PHP_70,
-        SetList::PHP_71,
         SetList::PHPUNIT,
     ]);
+
+    // add declare(strict_types=1); to all php files:
+    $ecsConfig->rule(DeclareStrictTypesFixer::class);
+
+    // change $array = array(); to $array = [];
+    $ecsConfig->ruleWithConfiguration(ArraySyntaxFixer::class, [
+        'syntax' => 'short',
+    ]);
+
+    // [default: PHP_EOL]; other options: "\n"
+    $ecsConfig->lineEnding("\n");
 };
